separate discover components, cache rt for single day
