+++ src/frontend/src/app/rotten-tomatoes/rotten-tomatoes.component.html
@@ -3,7 +3,7 @@
   <div class="card">
     <div class="card-header">Filters</div>
     <div class="card-body" *ngIf="form">
-      <form  [formGroup]="form">
+      <form [formGroup]="form">
         <input type="hidden" formControlName="page">
         <div class="row">
           <div class="col-12 col-md-6 col-lg-3">
